refactor: move IntermediateDecimalError into `proof-of-sql` crate
This error type will be tidied up in a future PR, but it is at least moved out of the `proof-of-sql-parser` crate.
Fixing errors types is a bit more complex, so to keep this PR small, we opt to only do this here, even though it is not tidy.
