--- crates/proof-of-sql-parser/src/intermediate_decimal.rs
@@ -1,37 +1,0 @@
-//! A parser conforming to standard postgreSQL to parse the precision and scale
-//! from a decimal token obtained from the lalrpop lexer. This module
-//! exists to resolve a cyclic dependency between proof-of-sql
-//! and proof-of-sql-parser.
-//!
-//! A decimal must have a decimal point. The lexer does not route
-//! whole integers to this contructor.
-use bigdecimal::ParseBigDecimalError;
-use snafu::Snafu;
-
-/// Errors related to the processing of decimal values in proof-of-sql
-#[allow(clippy::module_name_repetitions)]
-#[derive(Snafu, Debug, PartialEq)]
-pub enum IntermediateDecimalError {
-    /// Represents an error encountered during the parsing of a decimal string.
-    #[snafu(display("{error}"))]
-    ParseError {
-        /// The underlying error
-        error: ParseBigDecimalError,
-    },
-    /// Error occurs when this decimal cannot fit in a primitive.
-    #[snafu(display("Value out of range for target type"))]
-    OutOfRange,
-    /// Error occurs when this decimal cannot be losslessly cast into a primitive.
-    #[snafu(display("Fractional part of decimal is non-zero"))]
-    LossyCast,
-    /// Cannot cast this decimal to a big integer
-    #[snafu(display("Conversion to integer failed"))]
-    ConversionFailure,
-}
-impl From<ParseBigDecimalError> for IntermediateDecimalError {
-    fn from(value: ParseBigDecimalError) -> Self {
-        IntermediateDecimalError::ParseError { error: value }
-    }
-}
-
-impl Eq for IntermediateDecimalError {}
