+++ crates/proof-of-sql-parser/src/lib.rs
@@ -3,8 +3,6 @@
 #![cfg_attr(test, allow(clippy::missing_panics_doc))]
 extern crate alloc;
 
-/// Module for handling an intermediate decimal type received from the lexer.
-pub mod intermediate_decimal;
 /// Module for handling an intermediate timestamp type received from the lexer.
 pub mod posql_time;
 #[macro_use]
